1. Removed all metric plugin code from StatisticsHelper 2. Created MetricManager which houses this code 3. ??? 4. Profit
Minor tweaks to the MetricPlugin.hh; it now uses long unsigned int instead of uint32_t,
because that was the actual type being passed by some of the metrics in the Core apps.

MetricManager has state-change methods initialize, do_start, do_stop, do_pause, do_resume,
and shutdown(); these are called in the corresponding points in the *Core.cc files.

+++ artdaq/Plugins/MetricPlugin.hh
@@ -27,7 +27,7 @@ namespace artdaq {
   virtual void sendMetric(std::string name, int value, std::string unit) =0;
   virtual void sendMetric(std::string name, double value, std::string unit) =0;
   virtual void sendMetric(std::string name, float value, std::string unit) =0;
-  virtual void sendMetric(std::string name, uint32_t value, std::string unit) =0;
+  virtual void sendMetric(std::string name, long unsigned int value, std::string unit) =0;
   virtual void startMetrics() =0;
   virtual void stopMetrics() =0;
